split db logic
