+++ core/metrics/db.go
@@ -25,8 +25,6 @@ type entity struct {
 	Name  string `gorm:"column:name;uniqueIndex:entities_group_name"`
 }
 
-var ErrIncomplete = errors.New("meter profile incomplete")
-
 func init() {
 	db.Register(func(_ *gorm.DB) error {
 		return SetupSchema()
@@ -122,111 +120,6 @@ func SetupSchema() error {
 	return db.Instance.AutoMigrate(new(meter))
 }
 
-// Slot represents an aggregated energy time slot
-type Slot struct {
-	Start  time.Time `json:"start"`
-	End    time.Time `json:"end"`
-	Import float64   `json:"import"`
-	Export float64   `json:"export"`
-}
-
-// Series represents a named series of energy slots
-type Series struct {
-	Name  string `json:"name,omitempty"`
-	Group string `json:"group"`
-	Data  []Slot `json:"data"`
-}
-
-var aggregateFormats = map[string]string{
-	"15m":   "%Y-%m-%d %H:%M",
-	"hour":  "%Y-%m-%d %H:00",
-	"day":   "%Y-%m-%d",
-	"month": "%Y-%m",
-}
-
-var aggregateGoFormats = map[string]string{
-	"15m":   "2006-01-02 15:04",
-	"hour":  "2006-01-02 15:00",
-	"day":   "2006-01-02",
-	"month": "2006-01",
-}
-
-var aggregateDurations = map[string]func(time.Time) time.Time{
-	"15m":   func(t time.Time) time.Time { return t.Add(15 * time.Minute) },
-	"hour":  func(t time.Time) time.Time { return t.Add(time.Hour) },
-	"day":   func(t time.Time) time.Time { return t.AddDate(0, 0, 1) },
-	"month": func(t time.Time) time.Time { return t.AddDate(0, 1, 0) },
-}
-
-// QueryImportEnergy returns aggregated energy data, per entity or per group.
-func QueryImportEnergy(from, to time.Time, aggregate string, grouped bool) ([]Series, error) {
-	format, ok := aggregateFormats[aggregate]
-	if !ok {
-		return nil, errors.New("invalid aggregate value")
-	}
-
-	addDuration := aggregateDurations[aggregate]
-
-	groupCols := `e.name, e."group", bucket`
-	if grouped {
-		groupCols = `e."group", bucket`
-	}
-
-	type row struct {
-		Name   string
-		Group  string
-		Bucket string
-		Import float64
-		Export float64
-	}
-
-	tx := db.Instance.Table("meters m").
-		Select(`e.name, e."group", strftime(?, m.ts, 'unixepoch', 'localtime') AS bucket,
-			COALESCE(SUM(m."import"), 0) AS import, COALESCE(SUM(m.export), 0) AS export`, format).
-		Joins("JOIN entities e ON m.meter = e.id").
-		Group(groupCols).
-		Order(groupCols)
-
-	if !from.IsZero() {
-		tx = tx.Where("m.ts >= ?", from.Unix())
-	}
-	if !to.IsZero() {
-		tx = tx.Where("m.ts < ?", to.Unix())
-	}
-
-	var rows []row
-	if err := tx.Scan(&rows).Error; err != nil {
-		return nil, err
-	}
-
-	var res []Series
-	for _, r := range rows {
-		start, err := time.ParseInLocation(aggregateGoFormats[aggregate], r.Bucket, from.Location())
-		if err != nil {
-			return nil, err
-		}
-
-		name := r.Name
-		if grouped {
-			name = ""
-		}
-
-		if n := len(res); n == 0 || res[n-1].Name != name || res[n-1].Group != r.Group {
-			res = append(res, Series{Name: name, Group: r.Group})
-		}
-
-		s := &res[len(res)-1]
-		s.Data = append(s.Data, Slot{
-			Start:  start,
-			End:    addDuration(start),
-			Import: r.Import,
-			Export: r.Export,
-		})
-	}
-
-	return res, nil
-}
-
 // persist stores 15min consumption in kWh
 func persist(entity entity, ts time.Time, imp, exp float64) error {
 	return db.Instance.Create(&meter{
@@ -235,55 +128,4 @@ func persist(entity entity, ts time.Time, imp, exp float64) error {
 		Import:    imp,
 		Export:    exp,
 	}).Error
-}
-
-// importProfile returns a 15min average meter profile in Wh. The profile
-// is sorted by timestamp starting at 00:00. It is guaranteed to contain 96 15min values.
-func importProfile(entity entity, from time.Time) (*[96]float64, error) {
-	db, err := db.Instance.DB()
-	if err != nil {
-		return nil, err
-	}
-
-	rows, err := db.Query(`SELECT min(ts) AS ts, avg(import) AS import
-		FROM meters
-		WHERE meter = ? AND ts >= ?
-		GROUP BY strftime("%H:%M", ts, 'unixepoch', 'localtime')
-		ORDER BY strftime("%H:%M", ts, 'unixepoch', 'localtime') ASC`,
-		entity.Id, from.Unix(),
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var prev time.Time
-	res := make([]float64, 0, 96)
-
-	for rows.Next() {
-		var ts SqlTime
-		var val float64
-
-		if err := rows.Scan(&ts, &val); err != nil {
-			return nil, err
-		}
-
-		// interpolate single missing value, maybe due to regular restarts?
-		if time.Time(ts).Sub(prev) == 2*tariff.SlotDuration {
-			res = append(res, (val+res[len(res)-1])/2)
-		}
-		prev = time.Time(ts)
-
-		res = append(res, val)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-
-	if len(res) != 96 {
-		return nil, ErrIncomplete
-	}
-
-	return (*[96]float64)(res), nil
 }
